updated to current state
