+++ package.json
@@ -9,6 +9,7 @@
     "@fortawesome/react-fontawesome": "^0.1.7",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
+    "@material-ui/pickers": "^3.3.10",
     "@material-ui/styles": "^4.11.3",
     "@reduxjs/toolkit": "^1.2.1",
     "airtable-schema-generator": "1.5.0",
@@ -24,6 +25,7 @@
     "react": "^16.10.1",
     "react-app-rewire-alias": "^1.0.1",
     "react-app-rewired": "^2.1.8",
+    "react-calendar": "^3.4.0",
     "react-circular-progressbar": "^2.0.3",
     "react-dom": "^16.10.1",
     "react-modal": "^3.11.2",
